Only look for generic type if the field is generic
