+++ graphql-dgs/src/test/java/com/netflix/graphql/dgs/internal/java/test/inputobjects/JGenericBaseInputObject.java
@@ -16,24 +16,24 @@
 
 package com.netflix.graphql.dgs.internal.java.test.inputobjects;
 
-public class JGenericInputObject<A, B> {
+public class JGenericBaseInputObject<T> {
 
-    private A fieldA;
-    private B fieldB;
+    private String someField;
+    private T fieldA;
 
-    public A getFieldA() {
+    public String getSomeField() {
+        return someField;
+    }
+
+    public void setSomeField(String someField) {
+        this.someField = someField;
+    }
+
+    public T getFieldA() {
         return fieldA;
     }
 
-    public void setFieldA(A fieldA) {
+    public void setFieldA(T fieldA) {
         this.fieldA = fieldA;
-    }
-
-    public B getFieldB() {
-        return fieldB;
-    }
-
-    public void setFieldB(B fieldB) {
-        this.fieldB = fieldB;
     }
 }
